Bump package to 0.1.4
This has a breaking change - you're now required to pass a 3rd parameter to
`PrometheusMetrics::new` in order to account for custom labels.

As this is still a 0.Y.Z release, it's still within semver policy. Apologies if this
breaks code for someone. Once I ship a 1.Y.Z release, this will stop.